Reorganize code to allow for more modularity
- [FEATURE] Change management URL from / to /manage

- [FEATURE] Store api token in session storage

- [FEATURE] Add optional nav extension, customizable in a routes' meta

- [FEATURE] Added temporary screens for registration

- [REFACTOR] Seperate App.vue into Sidenav.vue, Nav.vue, Management.vue, Login.vue

- [REFACTOR] Move screen components to group folders

- [STYLE] Add trailing comma in prettier

- [FIX] Use v-main instead of main

- [CHORE] Update axios
