+++ .prettierrc
@@ -4,5 +4,7 @@
   "singleQuote": false,
   "endOfLine": "lf",
   "semi": true,
-  "vueIndentScriptAndStyle": false
+  "vueIndentScriptAndStyle": false,
+  "trailingComma": "all"
 }
+
