+++ src/components/Registration.vue
@@ -1,0 +1,13 @@
+<template>
+  <div>
+    <router-view></router-view>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "Registration",
+};
+</script>
+
+<style></style>
