+++ src/components/management/Delegates.vue
