+++ src/components/registration/FinalRegistration.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>Final Registration</div>
+</template>
+
+<script>
+export default {
+  name: "FinalRegistration",
+};
+</script>
+
+<style></style>
