+++ src/components/registration/Processing.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>Processing</div>
+</template>
+
+<script>
+export default {
+  name: "Processing",
+};
+</script>
+
+<style></style>
